Moved optimization to new module glotaran.optimization.
