+++ glotaran/builtin/megacomplexes/damped_oscillation/test/test_doas_model.py
@@ -3,12 +3,12 @@ from __future__ import annotations
 import numpy as np
 import pytest
 
-from glotaran.analysis.optimize import optimize
 from glotaran.builtin.megacomplexes.damped_oscillation import DampedOscillationMegacomplex
 from glotaran.builtin.megacomplexes.decay import DecayMegacomplex
 from glotaran.builtin.megacomplexes.spectral import SpectralMegacomplex
 from glotaran.model import Megacomplex
 from glotaran.model import Model
+from glotaran.optimization.optimize import optimize
 from glotaran.parameter import ParameterGroup
 from glotaran.project import Scheme
 from glotaran.simulation import simulate
