+++ glotaran/builtin/megacomplexes/decay/test/test_spectral_irf.py
@@ -5,9 +5,9 @@ from textwrap import dedent
 import numpy as np
 import pytest
 
-from glotaran.analysis.optimize import optimize
 from glotaran.io import load_model
 from glotaran.io import load_parameters
+from glotaran.optimization.optimize import optimize
 from glotaran.project import Scheme
 from glotaran.simulation import simulate
