+++ glotaran/builtin/megacomplexes/spectral/test/test_spectral_model.py
@@ -4,12 +4,12 @@ import numpy as np
 import pytest
 import xarray as xr
 
-from glotaran.analysis.optimize import optimize
-from glotaran.analysis.util import calculate_matrix
 from glotaran.builtin.megacomplexes.decay.test.test_decay_megacomplex import DecayModel
 from glotaran.builtin.megacomplexes.spectral import SpectralMegacomplex
 from glotaran.model import Megacomplex
 from glotaran.model import Model
+from glotaran.optimization.optimize import optimize
+from glotaran.optimization.util import calculate_matrix
 from glotaran.parameter import ParameterGroup
 from glotaran.project import Scheme
 from glotaran.simulation import simulate
