+++ glotaran/builtin/megacomplexes/test/test_spectral_decay_full_model.py
@@ -1,9 +1,9 @@
 import numpy as np
 import pytest
 
-from glotaran.analysis.optimize import optimize
 from glotaran.io import load_model
 from glotaran.io import load_parameters
+from glotaran.optimization.optimize import optimize
 from glotaran.project import Scheme
 from glotaran.simulation import simulate
