+++ glotaran/optimization/__init__.py
@@ -1,0 +1,1 @@
+"""This package contains functions for optimization."""
