+++ glotaran/optimization/test/test_grouping.py
@@ -1,8 +1,8 @@
 import numpy as np
 import xarray as xr
 
-from glotaran.analysis.optimization_group import OptimizationGroup
-from glotaran.analysis.test.models import SimpleTestModel
+from glotaran.optimization.optimization_group import OptimizationGroup
+from glotaran.optimization.test.models import SimpleTestModel
 from glotaran.parameter import ParameterGroup
 from glotaran.project import Scheme
