+++ glotaran/optimization/test/test_penalties.py
@@ -3,9 +3,9 @@ from copy import deepcopy
 import numpy as np
 import pytest
 
-from glotaran.analysis.optimization_group import OptimizationGroup
-from glotaran.analysis.test.models import TwoCompartmentDecay as suite
 from glotaran.model import EqualAreaPenalty
+from glotaran.optimization.optimization_group import OptimizationGroup
+from glotaran.optimization.test.models import TwoCompartmentDecay as suite
 from glotaran.parameter import ParameterGroup
 from glotaran.project import Scheme
 from glotaran.simulation import simulate
