+++ glotaran/optimization/util.py
